# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MemoryMesh API — AI Meeting Memory System powered by Cognee
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory/remember:
    post:
      operationId: rememberMeeting
      tags: [memory]
      summary: Cognee Remember — store and merge meeting into the memory graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RememberInput"
      responses:
        "200":
          description: Memory stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResult"
  /memory/recall:
    post:
      operationId: recallMemory
      tags: [memory]
      summary: Cognee Recall — retrieve relevant memories for a query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecallInput"
      responses:
        "200":
          description: Recalled memories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecallResult"
  /memory/improve:
    post:
      operationId: improveMemory
      tags: [memory]
      summary: Cognee Improve — update/evolve existing memories with new information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveInput"
      responses:
        "200":
          description: Memory improved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResult"
  /memory/forget:
    post:
      operationId: forgetMemory
      tags: [memory]
      summary: Cognee Forget — remove memories from the graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgetInput"
      responses:
        "200":
          description: Memories forgotten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
...
  /search:
    post:
      operationId: searchMemory
      tags: [search]
      summary: Semantic + graph hybrid search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchInput"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
  /search/suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [search]
      summary: Get real-time topic suggestions for a meeting
      parameters:
        - name: topic
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicSuggestion"

  # ─── AI CHAT ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RememberInput:
      type: object
      required: [meeting_id]
      properties:
        meeting_id:
          type: integer
        force_reindex:
          type: boolean
          default: false
    RecallInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        context_limit:
          type: integer
          default: 10
        meeting_ids:
          type: array
          items:
            type: integer
    ImproveInput:
      type: object
      required: [meeting_id, target_entity_id]
      properties:
        meeting_id:
          type: integer
        target_entity_id:
          type: integer
        relationship_type:
          type: string
          enum: [updates, supersedes, confirms, contradicts, refines]
    ForgetInput:
      type: object
      required: [scope]
      properties:
        scope:
          type: string
          enum: [meeting, project, entity, workspace]
        target_id:
          type: ["integer", "null"]
        reason:
          type: ["string", "null"]
    MemoryResult:
      type: object
      required: [success, entities_processed, relationships_created]
      properties:
        success:
          type: boolean
        entities_processed:
          type: integer
        relationships_created:
          type: integer
        merged_entities:
          type: integer
        message:
          type: string
    RecallResult:
      type: object
      required: [query, results, sources]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/RecallItem"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/MemorySource"
        total:
          type: integer
    RecallItem:
      type: object
      required: [content, relevance_score, entity_type]
      properties:
        content:
          type: string
        relevance_score:
          type: number
        entity_type:
          type: string
        entity_name:
          type: string
        meeting_id:
          type: ["integer", "null"]
        meeting_title:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
          format: date-time
    MemorySource:
      type: object
      required: [meeting_id, meeting_title, date, entity_type]
      properties:
        meeting_id:
          type: integer
        meeting_title:
          type: string
        date:
          type: string
          format: date-time
        entity_type:
          type: string
        entity_name:
          type: string
...
    OperationResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        affected_count:
          type: integer

    # ─── GRAPH ─────────────────────────────────────────────────────────────────
...
    SearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [semantic, graph, hybrid]
          default: hybrid
        entity_types:
          type: array
          items:
            type: string
        limit:
          type: integer
          default: 10
    SearchResult:
      type: object
      required: [query, results, total]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"
        total:
          type: integer
        search_mode:
          type: string
    SearchHit:
      type: object
      required: [id, title, type, score, excerpt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        score:
          type: number
        excerpt:
          type: string
        meeting_id:
          type: ["integer", "null"]
        meeting_title:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
          format: date-time
        entity_type:
          type: ["string", "null"]
    TopicSuggestion:
      type: object
      required: [topic, related_meetings, last_discussed]
      properties:
        topic:
          type: string
        related_meetings:
          type: array
          items:
            type: string
        last_discussed:
          type: string
          format: date-time
        conflict_warning:
          type: ["string", "null"]
        meeting_count:
          type: integer

    # ─── CHAT ──────────────────────────────────────────────────────────────────
...